requirements.txt
asgiref==3.3.1
click==7.1.2
cycler==0.10.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dlib==19.21.1
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.1
openpyxl==3.0.7
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
XlsxWriter==1.3.8
